src/requirements.txt
djangorestframework==3.12.2
drf-yasg==1.20.0
django-oauth-toolkit==1.4.0
